moved bits, bytes, mem interface into utils, moved timers into common
