+++ spa/src/utils/bits.rs
