+++ spa/src/utils/bytes.rs
