+++ spa/src/utils/meminterface.rs
@@ -10,25 +10,25 @@ pub trait MemInterface8 {
     fn write_byte(&mut self, addr: u32, data: u8);
 
     fn read_halfword(&mut self, addr: u32) -> u16 {
-        use crate::common::bytes::u16;
+        use crate::utils::bytes::u16;
         let lo = self.read_byte(addr);
         let hi = self.read_byte(addr + 1);
         u16::make(hi, lo)
     }
     fn write_halfword(&mut self, addr: u32, data: u16) {
-        use crate::common::bytes::u16;
+        use crate::utils::bytes::u16;
         self.write_byte(addr, u16::lo(data));
         self.write_byte(addr + 1, u16::hi(data));
     }
 
     fn read_word(&mut self, addr: u32) -> u32 {
-        use crate::common::bytes::u32;
+        use crate::utils::bytes::u32;
         let lo = self.read_halfword(addr);
         let hi = self.read_halfword(addr + 2);
         u32::make(hi, lo)
     }
     fn write_word(&mut self, addr: u32, data: u32) {
-        use crate::common::bytes::u32;
+        use crate::utils::bytes::u32;
         self.write_halfword(addr, u32::lo(data));
         self.write_halfword(addr + 2, u32::hi(data));
     }
@@ -41,7 +41,7 @@ pub trait MemInterface8 {
 /// Lower bytes will be read/written first.
 pub trait MemInterface16 {
     fn read_byte(&mut self, addr: u32) -> u8 {
-        use crate::common::bytes::u16;
+        use crate::utils::bytes::u16;
         let data = self.read_halfword(addr & 0xFFFF_FFFE);
         match addr & 1 {
             0 => u16::lo(data),
@@ -50,7 +50,7 @@ pub trait MemInterface16 {
         }
     }
     fn write_byte(&mut self, addr: u32, data: u8) {
-        use crate::common::bytes::u16;
+        use crate::utils::bytes::u16;
         let halfword_addr = addr & 0xFFFF_FFFE;
         let halfword_data = self.read_halfword(halfword_addr);
         match addr & 1 {
@@ -64,13 +64,13 @@ pub trait MemInterface16 {
     fn write_halfword(&mut self, addr: u32, data: u16);
 
     fn read_word(&mut self, addr: u32) -> u32 {
-        use crate::common::bytes::u32;
+        use crate::utils::bytes::u32;
         let lo = self.read_halfword(addr);
         let hi = self.read_halfword(addr + 2);
         u32::make(hi, lo)
     }
     fn write_word(&mut self, addr: u32, data: u32) {
-        use crate::common::bytes::u32;
+        use crate::utils::bytes::u32;
         self.write_halfword(addr, u32::lo(data));
         self.write_halfword(addr + 2, u32::hi(data));
     }
