+++ spa/src/utils/mod.rs
@@ -1,0 +1,5 @@
+// Shared helper functions.
+
+pub mod bits;
+pub mod bytes;
+pub mod meminterface;
